API/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Date/Time
python-dateutil==2.9.0

# Excel/PDF export
openpyxl==3.1.5
reportlab==4.2.4
xlsxwriter==3.2.0

# HTTP client
requests==2.32.3
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
